+++ webgraph/src/graphs/arc_list_graph.rs
@@ -71,8 +71,8 @@ where
     where
         Self: 'a;
 
-    fn split_iter(&self, how_many: usize) -> Self::IntoIterator<'_> {
-        split::seq::Iter::new(self.iter(), self.num_nodes(), how_many)
+    fn split_iter_at(&self, cutpoints: impl IntoIterator<Item = usize>) -> Self::IntoIterator<'_> {
+        split::seq::Iter::new(self.iter(), cutpoints)
     }
 }