+++ webgraph/src/graphs/btree_graph.rs
@@ -224,8 +224,8 @@ impl<L: Clone + Sync> SplitLabeling for LabeledBTreeGraph<L> {
     where
         Self: 'a;
 
-    fn split_iter(&self, how_many: usize) -> Self::IntoIterator<'_> {
-        split::ra::Iter::new(self, how_many)
+    fn split_iter_at(&self, cutpoints: impl IntoIterator<Item = usize>) -> Self::IntoIterator<'_> {
+        split::ra::Iter::new(self, cutpoints)
     }
 }
 
@@ -453,8 +453,8 @@ impl SplitLabeling for BTreeGraph {
     where
         Self: 'a;
 
-    fn split_iter(&self, how_many: usize) -> Self::IntoIterator<'_> {
-        split::ra::Iter::new(self, how_many)
+    fn split_iter_at(&self, cutpoints: impl IntoIterator<Item = usize>) -> Self::IntoIterator<'_> {
+        split::ra::Iter::new(self, cutpoints)
     }
 }