+++ webgraph/src/graphs/vec_graph.rs
@@ -345,8 +345,8 @@ impl<L: Clone + Sync> SplitLabeling for LabeledVecGraph<L> {
     where
         Self: 'a;
 
-    fn split_iter(&self, how_many: usize) -> Self::IntoIterator<'_> {
-        split::ra::Iter::new(self, how_many)
+    fn split_iter_at(&self, cutpoints: impl IntoIterator<Item = usize>) -> Self::IntoIterator<'_> {
+        split::ra::Iter::new(self, cutpoints)
     }
 }
 
@@ -611,8 +611,8 @@ impl SplitLabeling for VecGraph {
     where
         Self: 'a;
 
-    fn split_iter(&self, how_many: usize) -> Self::IntoIterator<'_> {
-        split::ra::Iter::new(self, how_many)
+    fn split_iter_at(&self, cutpoints: impl IntoIterator<Item = usize>) -> Self::IntoIterator<'_> {
+        split::ra::Iter::new(self, cutpoints)
     }
 }