+++ webgraph/src/traits/graph.rs
@@ -307,8 +307,8 @@ where
     where
         Self: 'a;
 
-    fn split_iter(&self, how_many: usize) -> Self::IntoIterator<'_> {
-        self.0.split_iter(how_many).into_iter().map(UnitLender)
+    fn split_iter_at(&self, cutpoints: impl IntoIterator<Item = usize>) -> Self::IntoIterator<'_> {
+        self.0.split_iter_at(cutpoints).into_iter().map(UnitLender)
     }
 }